==> R/L1/4_clean_network_data.qmd <==
---
title: "AvianMetaNet: L1 taxonomic corrections and updates"
author: 
  - Phoebe L. Zarnetske
  - Patrick Bills
  - Kelly Kapsar
format:
  html:
    toc: true
    toc-depth: 2
    embed-resources: true
editor: source
date: 2026-02-06
date-format: iso
---

```{r}
#| echo: false
#| output: false
# Clear all existing data
rm(list=ls())
```

## About

- PROJECT: AvianMetaNet
- COLLABORATORS: Vincent Miele, Stephane Dray
- Run Date: `r Sys.Date()`
- Next script to run: 5_subset_network.qmd
  this cleans further (e.g., breeding/non-breeding) and some of that needs to go in general cleaning (either here or in another script).

This script is used to refine species name changes to align with the 
Clements & eBird checklist to create an Rdata file.  

Currently, the script is working for a checklist of CONUS+Alaska+Canada.  

L0 data are checked to assign updated scientific and common names to the interaction pairs data. 

### Workflow Summary

At each step, resolution narrows down those names that require manual entry 
in a table to identify the current accepted name in checklist

1. setup 
   - read in checklists and L0 data (int.raw)
   - final data cleaning to ensure no typos etc
   - create a name resolution table to be completed (int.raw.names)
2. fix missing scientific names manually if any (only 1)
3. resolve scientific names already exactly matching checklist 
4. for those names not resolved, use fuzzy matching to find scientific names in checklist (high confidence only)
5. for those names not resolved, match common names with checklist (high confidence only)
6. for those names not resolved, use GBIF to see if name can be resolved to a checklist name
7. read in CSV of manual resolutions and apply to whole list 
8. resolve genus only unids
9. resolve hybrids ( X and / )
10. merge checklist common names into resolutions matching on scientific name
11. join checklist scientific names and common names to interactions data
12. save data files 
    - interaction data with checklist names resolved 
    - resolutions table with explanations

### Major Revisions

- Original R script `AvianInteractionData_L1.R` 27 Oct 2022 (PLZ)
- updated through 12 Dec 2024 (PLZ);
  -   As of Dec. 12, 2024, this is mainly cleaned for BBS species
- May 2025- July 2025 convert to workflow notebook/Quarto, (PSB)
- August 2025, using new species lists Canada+CONUS from PLZ (PSB)
- September 10, 2025, using species lists from commit from PLZ (PSB)
- Oct 1, 2025: workflow reorganization for better taxonomic resolution (PSB; PLZ edits and checks)
- Nov 2025- January 2026: rename and edits to better system workflow organization, variable names, 
           handling UniD (genus-level) entries (KK)
- Feb 6, 2026: Quarto notebook edits for readable data in rendered report (PSB)


## Setup

### Data inputs:

#### **AvianMetaNet L0**

Raw Data, combined (stitched) and with content
cleaning but taxonomic names as entered (older names, typos, etc)

*Set the interactions database file to read from here: ( output From `R/L0/2_stitch_species.R` )*

```{r}
AvianInteractionData_L0_file <-  "amn_all_raw.csv"
```
#### **Main Checklists:**

- Full 2024 Clements/eBird list 
- GBIF database (via taxadb package)

Curated Species checklist: 

- Current selected species list =
  combined checklist based on Clements/eBird v2024, Avibase v8.17, BBS v2024:
  `R/L0/3_subset_species_lists.R`
  - used for species lookup / matching

#### **Manual Taxonomic Resolution Overrides Table:**

- Rows of corrections to be applied to L0 data for updating species 
as entered to match the current checklist.  Update the code below 

```{r}

manual_taxonomic_resolutions_file_name<- here::here('R/L1', "aux_taxonomy_resolutions_202510.csv")
clements_mismatches <- readr::read_csv(here::here("./data/L1/aux_mismatches_with_clements_best_match.csv"))

```

### Data Outputs

#### Taxonomic Resolution Table: 

AvianMetaNet, cleaned and with taxonomic names resolved to 
match checklists listed above: `amn_all.csv`

```{r}
AvianInteractionData_L1_file <- "amn_all.csv" 
AvianInteractionData_L1_taxonomy_table_file <- "aux_taxonomy_resolutions.csv"

```

### Setup: File paths used for this run:

```{r}
#| echo: false

# functions used by all 
source(here::here('R/lib/shared_functions.R'))
# functions for this process to reduce this script size
# and load libraries used here
source(here::here('R/lib/taxonomy_functions.R'))

file_paths <- get_file_paths()

print(file_paths)
```

*Note that the Checklist folder above is in L0 which contains species lists, 
but the subset list for 2024-25 analysis is in L1 as it's built from L0 lists*

### Setup: Checklists

#### Curated List

**Previously:** all species in BBS `"bbs_splist_2024_L1.csv"`

(the 2024 BBS release includes all species as of 2023 BBS surveys, plus the additional 
AOUcombo.index column (= lumps BBS subspecies into species)for use w BBS AOU index)

`splist2024<-read.csv(file.path(L1_dir,"bbs_splist_2024_L1.csv"))`

**Current**

September 2025: new curated list curated by PLZ using Clements, AVIBase, BBS etc. 
See `R/L1/3_subset_species_lists.R`

```{r}

current_species_list_filename <- "spp_joint_cac.csv"
splist2024_filepath <- file.path(file_paths$CHECKLIST_L1,current_species_list_filename )
splist2024 <- read.csv(here::here(splist2024_filepath))
print(paste("Read in ", nrow(splist2024), "entries from checklist file", splist2024_filepath))

```

#### Main Checklist

Read in the full check list the official eBird/Clements v2024 checklist

```{r}
#| label: read_clements


checklist_filename <- "eBird-Clements-v2024-integrated-checklist-October-2024-rev.csv"
checklist.orig <- read.csv(here::here(file.path(file_paths$CHECKLIST_L0,checklist_filename)))
print(paste("Read in ", nrow(checklist.orig), "entries from checklist file", checklist_filename))
```

amend column names of checklist to match for merging later show col name before and after change

```{r}
print(names(checklist.orig))
checklist <- checklist.orig
names(checklist)[names(checklist) == "scientific.name"] <-"scientific_name"
names(checklist)[names(checklist) == "English.name"] <-"common_name"
print("CORRECTED CHECKLIST COLUMNS:")
print(names(checklist))
```

set how we find 'species' in checklist

*The Clements checklist includes families, species, hybrids, etc.   In several
places below we want to find only the 'species' which include a few 
categories.  This variable is used in several places in this notebook
to search only 'species' in the checklist*

```{r}
species_categories <- c("group (monotypic)","species","slash","subspecies","genus","family")
```

```{r}
#| echo: false

# cheap function using global var checklist to make it easy
on_checklist <- function(sciname){
  rows<- filter(checklist, scientific_name == sciname)
  if(nrow(rows)>0 ) return(TRUE)
  
  return(FALSE)
}
```

### Interactions Data Set-up

To use a different or test files, edit this L0 filename:

```{r}
#| echo: false
full_path_stitched_file <- file.path(file_paths$L0,AvianInteractionData_L0_file)
int.raw<-read.csv(full_path_stitched_file)
print(paste("read in ", nrow(int.raw), "records from ", full_path_stitched_file))
```

## Final Interaction Data Cleaning/Corrections

Most of these filters and adjustments are now made during the 2_stitch_species process
but were are ensuring the data is clean prior to matching taxonomy


#### Remove blank rows

```{r}
before_filter_nrow <- nrow(int.raw)
int.raw <- filter(int.raw,!(is.na(taxa1_scientific)& 
                             is.na(taxa2_scientific) & 
                             is.na(taxa1_common) & 
                             is.na(taxa2_common)
                           ))

print( paste(as.integer(before_filter_nrow - nrow(int.raw)), "blank rows removed"))
```

#### Trim white space from all names columns

```{r}
# trim white space

int.raw <- dplyr::mutate(int.raw, 
                         taxa1_scientific = trimws(taxa1_scientific,which=c("right")), 
                         taxa2_scientific = trimws(taxa2_scientific,which=c("right")), 
                         taxa1_common = trimws(taxa1_common,which=c("right")),
                         taxa2_common = trimws(taxa2_common,which=c("right"))
) 

print("unique scientific name combos after removing white space")
print(length(unique(paste(int.raw$taxa1_scientific, int.raw$taxa2_scientific))))
```

#### Missing Scientific name check

Are there any records that are missing a scientific name for species1?

*if so this is problematic for adjusting the recorded name with the checklist 
name, and should be corrected in data entry files*

```{r}
records.missingsp1 <- int.raw[is.na(int.raw$taxa1_scientific),]
tibble(records.missingsp1)
```

Records missing a scientific name for sp2

```{r}
records.missingsp2 <- int.raw[is.na(int.raw$taxa2_scientific),]
tibble(records.missingsp2)

```

Are there rows that have *only* common names for both sp1 and sp2?

```{r}
 
int.common_names_only <- dplyr::filter(int.raw, is.na(taxa1_scientific) & is.na(taxa2_scientific))
  
print(nrow(int.common_names_only))
```

#### remove rows that do not have interactions

*These should be corrected in the L0 correct process, 
but catch them here before correcting taxonomy*

List rows with both interaction columns missing

```{r}
#| echo: false
# to allow us to re-run this, look to see if we saved the var with all rows
# because we've run the cells below, and if so, reset our main df to that one
if(exists("int.raw.saved_all_rows")) int.raw<- int.raw.saved_all_rows

print("n rows with both effects missing:")
print(nrow(filter(int.raw, is.na(effect_tx1_on_tx2) & is.na(effect_tx2_on_tx1))))
print("n rows with at least one effect present:")
print(nrow(filter(int.raw, !is.na(effect_tx1_on_tx2) | !is.na(effect_tx2_on_tx1))))
print("list of rows with both effects missing:")
```

Table of rows with both effects missing:

```{r}
#| echo: false

tibble(filter(int.raw, is.na(effect_tx1_on_tx2) & is.na(effect_tx2_on_tx1)))
```

Remove those with no interactions (if any)
```{r}

# save the original full list in case we need to check again
int.raw.saved_all_rows <- int.raw

int.raw <- filter(int.raw, !is.na(effect_tx1_on_tx2) | !is.na(effect_tx2_on_tx1))
print("rows remaining with not both effects NA:")
print(nrow(int.raw))
```

#### Discovery and repair of missing Sp1/Sp2 scientific names

List rows with missing sp1 values but have other data to use for fix

```{r}
records.missingsp1 <- int.raw[is.na(int.raw$taxa1_scientific),]
print(records.missingsp1)
```

Fix these missing species directly by matching the common_names from checklist

```{r}
#| echo: false
 
for(common_name_for_missingsp1  in records.missingsp1$taxa1_common){
  print(common_name_for_missingsp1)
  
  matching_sp<- checklist %>% 
    dplyr::filter(common_name == common_name_for_missingsp1) %>% 
    dplyr::select(common_name, scientific_name)  
  
  print(as.vector(matching_sp[1,]$scientific_name))
  # print(subset(checklist, common_name == common_name_for_missingsp1))
}
```


Assign scientific names to all rows for those common names that have missing sp1 scientific names, create a Manual Assignment structure here:

```{r}
#| echo: false
manual_assignments = list()
manual_assignments[["Rhinoceros Auklet"]] <- "Cerorhinca monocerata"

print(manual_assignments)
# add more as needed  
```

apply the manual assignments

```{r}
for(cname in names(manual_assignments)){
  print(paste("set records with common name", cname, "to sp1 = ", manual_assignments[[cname]]))
  print("n rows to update:")
  print(nrow(int.raw[int.raw$taxa1_common == cname & is.na(int.raw$taxa1_scientific),]))
  int.raw[
    int.raw$taxa1_common == cname & 
    is.na(int.raw$taxa1_scientific),
    ]$taxa1_scientific = manual_assignments[[cname]] 
}

  
```

double check: remaining records with **missing sp1:**

```{r}
records.missingsp1 <- int.raw[is.na(int.raw$taxa1_scientific),]
print(records.missingsp1)  
```

### Discovery and repair of missing Sp2 scientific

rows with missing **sp2 scientific**:

```{r}
records.missingsp2<- int.raw[is.na(int.raw$taxa2_scientific),]
print(records.missingsp2)
```


Saved outcome from Previous run
```         
May 2025: 2 rows need editing 
taxa1_common taxa2_common         taxa1_scientific
866         White-cheeked Pintail   Cinnamon Teal Anas bahamensis rubirostris
14223 Middle American Screech-Owl            <NA>        Megascops guatamalae
taxa2_scientific effect_tx1_on_tx2 effect_tx2_on_tx1          interaction
866                  <NA>                 1                 1 facilitation-feeding
14223                <NA>                NA                NA                 <NA>
one had both NA interactions , now we are removing those
```



List the names from the checklist that match those missing in sp2:

```{r}
#| echo: false
#| 
for(common_name_for_missingsp2  in records.missingsp2$taxa2_common){
  print(checklist %>% 
    dplyr::filter(common_name == common_name_for_missingsp2) %>% 
    dplyr::select(common_name, scientific_name))
}
```

**Sp2 fix list**

Update this list of fixes as errors are discovered above.

Note that the common names will be update updated below with the checklist.

```{r}
#| echo: false
manual_assignments = list()
manual_assignments["Cinnamon Teal"] <-  "Spatula cyanoptera"
# add more as needed  
print(manual_assignments)
```

fix sp2 using the list above

```{r}
# species2 fixes using the list above
for(cname in names(manual_assignments)){
  print(paste("set records with common name", cname, "to sp2 = ", manual_assignments[[cname]]))
  print("n rows to update with this command name and missing sp2:")
  rows_to_fix <- nrow(
    int.raw[int.raw$taxa2_common == cname & is.na(int.raw$taxa2_scientific),]
  )
  
  print(rows_to_fix)
  if(rows_to_fix > 0 ){
  int.raw[
    int.raw$taxa2_common == cname & 
    is.na(int.raw$taxa2_scientific),
    ]$taxa2_scientific = manual_assignments[[cname]] 
  }
}
  # double check remaining rows with missing sp2 (should be zero)
```

Double check that there are no remaining rows with missing scientific names

```{r}
tibble(int.raw[is.na(int.raw$taxa2_scientific),]) 
```
*completed fixes for missing scientific names*





2. ## Taxonomy Adjustments Table

**Create new data frame of all names, as is in the raw file, and what is to be 
adjusted/updated/assigned.


Column Definitions:

- scientific_name.raw : L0 name as it appears in the original CSV file (stitched)
- common_name.raw : L0 name as it appears in the original CSV file (stitched)
- scientific_name_clements2024 : L1 name to be used (pre checklist )
- common_name_clements2024 : L1 name to be used (pre checklist )
- edit_notes: text of why edits were made

The edit columns are based on the Clements 2024 checklist.
Determine how many unique entries are in int.raw\$taxa1_scientific, which 
are unique combos of scientific names.

```{r}
unique.sp.raw <- data.frame(scientific_name.raw = with(int.raw, union(
  taxa1_scientific, taxa2_scientific
)))

print(dim(unique.sp.raw)) #3913 # May 29, 2025: 3889; Sept. 29, 2025: 3879; Oct 26: 3873
```


**Create Name Edits Table**

Create list of unique scientific_name.raw and common_name.raw pairs with 
formatting adjustments, stacking and collecting anything in sp1 and sp2. 
See documentation in `taxonomy_functions.R` script for details

*note:  since adjustments/corrections to species were added to L0 cleaning 
process, there should be very few differences between 'raw' and 'edited' 
names at this point in the process*

```{r}
#| echo: false
# create the raw <- edit adjustments table 
int.raw.names <- create_names_edit_table(int.raw)
head(tibble(int.raw.names))
```

Number of unique species in taxonomy edits table  - no resolutions made yet

This May include entries that are duplicates but with a typo or old species name that needs resolution. 

```{r}
#| echo: false
#| output: true

count_L0_total_unique_species_unedited<- nrow(unique(unresolved_species(int.raw.names)))

```

Historical numbers of total species from L0 data:
Feb 2026: 3873  
May 2025: 3889
Sept 2025: 3876
Oct 2025: 3876


Test case to track with known misspelling: Accipiter cooperi was misspelled (should be Accipiter cooperii) - exists in int.raw.names

```{r}
known_misspellings <- c("Accipiter cooperi")
tibble(dplyr::filter(int.raw.names, scientific_name.raw %in% known_misspellings))
```

expected output:

```         
scientific_name.raw common_name.raw scientific_name.edit common_name.edit
1 Accipiter cooperi   Cooper's Hawk  Accipiter cooperi    Cooper's Hawk
```


**Example Checking process:**


Discovery that Dendroica pinus - Pine Warbler - genus was updated to Setophaga pinus
checklists for the old and then new name FOR THIS SPECIES

```{r}
scientific_name.raw <- "Dendroica pinus"

```

is this is being used? records with old name: 
```{r}
#| echo: false
#| output: true
print(interaction_count_by_species(int.raw,scientific_name.raw  ))

```
records with new name:
```{r}
#| echo: false
#| output: true

print(interaction_count_by_species(int.raw, "Setophaga pinus"))
```

Which is preferred on the checklist?
```{r}
checklist[which(checklist$scientific_name == "Dendroica pinus"), ] # no records
checklist[which(checklist$scientific_name == "Setophaga pinus"), ] # one record
```

Is there an entry in our name edit table with the old genus?
```{r}
tibble(int.raw.names[int.raw.names$scientific_name.raw == "Dendroica pinus",])
```

add edited name and note

```{r}
int.raw.names <- add_name_edits(edits.df = int.raw.names,
               scientific_name.raw = "Dendroica pinus", 
               scientific_name.edit = "Setophaga pinus",
               edit_notes = "Update to new checklist name"
)
```




## Checklist matches

Identify those species from the database that are already on the latest checklist
and add a note to track those. These do not need to be resolved. 

```{r}

# update in place

on_checklist_note <- "L0 species as entered on checklist (Clements 2024)"

checklist.species <- checklist[checklist$category %in% species_categories, "scientific_name"]

# make a copy
int.raw.names.checklist_match <- int.raw.names
# count them
count_L0_rows_scinames_already_on_checklist<- nrow(int.raw.names.checklist_match[int.raw.names$scientific_name.edit %in% checklist.species, ])
# add note when it matches
count_L0_species_scinames_already_on_checklist <- length(unique(int.raw.names.checklist_match[int.raw.names$scientific_name.edit %in% checklist.species,]$scientific_name.raw))

int.raw.names.checklist_match[int.raw.names$scientific_name.edit %in% checklist.species, ]$edit_notes <- on_checklist_note

# optional, review all those taxa that are not found in Clements
# print(unresolved_species(int.raw.names.checklist_match))

```

### L0 Checklist matching Summary

Number rows with species already on the checklist, so already resolved, which includes
full species and subspecies, but not slashes, unids, families or hybrids

```{r}
#| echo: false
#| output: true

nrow(filter(int.raw.names.checklist_match, !is.na(edit_notes) ))
```

<!-- **THIS IS INCORRECT**

Unique scientific names in the database NOT matched to the checklist : -->
```{r}

# nrow(unique ( unresolved_species(int.raw.names.checklist_match)))

```
September 2025: 545 rows 
October 5, 2025: 429 rows


save to actual list and continue

```{r}
int.raw.names <- int.raw.names.checklist_match
```

Remaining (unique) scientific names to resolve: 

```{r}
#| echo: false
#| output: true

count_scinames_needing_to_be_matched <- length(unique(unresolved_species(int.raw.names)$scientific_name.raw))
print(count_scinames_needing_to_be_matched)

```
February 2024: 425
September 2025: 504 (just about 30 redundant common names)
October 6, 2025: 404
February 2026: 401


Remaining Rows (sciname + common name combo) that need checking
```{r}
#| echo: false
#| output: true

count_rows_needing_matching_after_clements_match <- length(unique(unresolved_species(int.raw.names)$scientific_name.raw))

print(count_rows_needing_matching_after_clements_match )
```




## Fuzzy Matching of Scientific Name to checklist

Use fuzzy logic function ( see taxonomy_functions.R ) to find closest match from the CHECKLIST reference list
Function to find the closest match with a similarity score, try and determine what misspellings exist 

#### Reference list of scientific names from eBird Clements CHECKLIST 2024**

Create a reference list from our Checklist read in above (Clements)

```{r}


# checklist.species<- filter(checklist, category == "species" | category == "slash" | category == "subspecies")
checklist.species<- filter(checklist, category %in% species_categories)

reference_names <- tibble(
  scientific_name = checklist.species$scientific_name,
  common_name = checklist.species$common_name
)
nrow(reference_names)

```
#### Number of names to resolve: 

```{r}

count_scinames_unresolved_prior_to_sciname_fuzzy_matching <- length(unique(unresolved_species(int.raw.names)$scientific_name.edit))
print(count_scinames_unresolved_prior_to_sciname_fuzzy_matching)
```
### Resolve matches based on scientific_name.edit and Clements Checklist

Calculate the fuzzy match score and closest name in the checklist for 
each record in the unresolved names

*To examine the matches created by this process use `View(scientific_name_matches)`*

*(this takes several minutes to complete)*

```{r}
#| label: sciname_fuzzy_match_loop

# prepare a data frame to fill up with match data with current unresolved species
scientific_name_matches <- unresolved_species(int.raw.names)
# add new columns
# tracking original row number will make subsetting and cross-ref easier later
scientific_name_matches$row_number <- unresolved_species_rows(int.raw.names)
scientific_name_matches$closest_scientific_name_match <- NA
scientific_name_matches$scientific_name_match_score <- NA


# fill them in one row at a time
for(r in 1:nrow(scientific_name_matches)) {
  # optional, print the name we are on to gauge progress
  # print(scientific_name_matches[r,"scientific_name.edit"])
  fuzzy_match <- find_closest_match_with_score( scientific_name_matches[r,"scientific_name.edit"], checklist$scientific_name)
  scientific_name_matches[r,]$closest_scientific_name_match <- fuzzy_match$match
  scientific_name_matches[r,]$scientific_name_match_score <- fuzzy_match$score
}

# idea to improve speed: rewrite scientific_name_matches to accept a vector/list of names 
# and return list of lists, then rbind/cbind

tibble(scientific_name_matches)
```

Distribution of Match Scores

```{r}
#| label: fuzzy_match_historgram
#| echo: false
#| output: true


suppressMessages(suppressWarnings(library(ggplot2))) # don't print to console
g<- ggplot(scientific_name_matches, aes(scientific_name_match_score)) +  
        geom_histogram(binwidth=0.01, color = "#000000", fill = "#0099F8") # + 

print(g)

```


Test of checklist and data : Accipiter cooperi is misspelled AND has been changed to Astur, 
so the match is not even the same  - should have a match of about 0.84

```{r}

incorrect_sp_to_check <-"Accipiter cooperi" 
print(dplyr::filter(scientific_name_matches, 
                    scientific_name.raw %in% c(incorrect_sp_to_check))$scientific_name_match_score)
```



#### Scientific Name Matching Set High Confidence threshold

To review fuzzy match results on unresolved names, run `View(scientific_name_matches)`

PLZ Oct 5: after reviewing, and ordering from 1.0 to 0, found that Cercotrichas coryphoeus
Karoo Scrub-Robin (0.92555990) was the first that did not match.

After review of matches, all of the matches with score > 0.95 looked correct, 
and many but not all in range from 0.095 > score > 0.90 so those need to be checked. 
PLZ Oct 5: after reviewing, and ordering from 1.0 to 0, found that Cercotrichas coryphoeus
Karoo Scrub-Robin (0.92555990) was the first that did not match because Genus changed (to Tychaedon).

Setting the threshold to 0.95 for automatic assignment, and will add common name 
matching to increase confidence in the others.

```{r}
high_confidence_threshold <- 0.95
```

Number of unique species with high confidence matches (scientific_name_match_score >= `{r} high_confidence_threshold` )

```{r}
#| echo: false
#| output: true
high_confidence_matches <-  filter(scientific_name_matches, 
                                   scientific_name_match_score >= high_confidence_threshold
                                   )
print(length(unique(high_confidence_matches$scientific_name.raw)))
proportion_high_confidence = length(unique(high_confidence_matches$scientific_name.raw)) / length(unique(scientific_name_matches$scientific_name.raw))
print(sprintf("%.2f%%", round(proportion_high_confidence*100,2)))
```
*previous results:*

`Dec 2024: 257 out of 326; May 29, 2025: 244 out of 309`

Sept 2025:
  - >= 0.95 134 (33%)
  - >= 0.90 211
Oct 2025:
  - >= 0.93 166 (41.09%)
Oct 26, 2025:
  - >= 0.95 134 (33.17%) - same as Sept
  
### Resolve (assing edit_note to all high confidence matches)

```{r}

edit_note <- "high confidence str match to checklist (Clements 2024) "
# make a copy
int.raw.names.matches <- int.raw.names
# fill the matching rows with the matched sci names
# this assignment works becuase the match data frame contains the original row numbers
int.raw.names.matches[high_confidence_matches$row_number,]$scientific_name.edit<- high_confidence_matches$closest_scientific_name_match
int.raw.names.matches[high_confidence_matches$row_number,]$edit_notes <- paste( edit_note, round(high_confidence_matches$scientific_name_match_score,3))

# to review these assignments, run 
#   tibble(int.raw.names.matches) 
# which includes all resolutions
```

Remaining species to resolve after high-confidence scientific name match:

```{r}
#| echo: false
#| output: true
print("unresolved rows in names table:")
print(length(unresolved_species_rows(int.raw.names.matches)))
print("unresolved unique scientific names ")
print(length(unique(unresolved_species(int.raw.names.matches)$scientific_name.raw)))

```

```{r}
#| echo: false
# after review, save these resolutions back to main list variable
int.raw.names <- int.raw.names.matches
rm(int.raw.names.matches)
```



### Fuzzy match lower confidence matching


```{r}

medium_confidence_threshold <- 0.90
minimum_confidence_threshold <- 0.80

medium_confidence_matches <- filter(scientific_name_matches, 
                                     scientific_name_match_score < high_confidence_threshold & 
                                     scientific_name_match_score >= medium_confidence_threshold)

low_confidence_matches <- filter(scientific_name_matches, 
                                     scientific_name_match_score < medium_confidence_threshold & 
                                     scientific_name_match_score >= minimum_confidence_threshold)

```

### Determine lower confidence threshold

After review it appears that names may be matched correctly at a lower threshold
than some names matched incorrectly (at higher threshold), that is, it's a mix
from in matches from 0.90 - 0.95. 

Hold these and match on common name first, then re-run fuzzy matching

skip this because the histogram above shows this information. 
```{r}
# score_start <- min(high_confidence_matches$scientific_name_match_score)
# score_end <- max(high_confidence_matches$scientific_name_match_score)
# increment <- 0.005
# 
# # collect these in a list for exploring
# confidence_partition <- list()
# 
# # Loop through each score range and print the matches within that range
# for (i in seq(score_start, score_end, by = increment)) {
#   current_range <- high_confidence_matches %>%
#     filter(scientific_name_match_score >= i & scientific_name_match_score <= i + increment)
#   
#   # Print the current range if it has any entries
#   if (nrow(current_range) > 0) {
#     low_range <- sprintf("%.2f", i)
#     high_range <- sprintf("%.2f", i + increment)
#     range_text = paste(low_range, "to", high_range)
#     print(paste("Match Score Range:",range_text ,": ", nrow(current_range), "records"))
#     confidence_partition[[high_range]]<- current_range
#     # print(current_range, n=100) # print up to 100 rows in a section
#   }
# }
# 
# ```
# View any of these dataframes using the partition list to see what's on there
# ```{r}
#  confidence_partition[["0.91"]]
```


```{r}
#| echo: false

# **optional CHECKPOINT: save current workspace with current date**
# today_date_string <- format(Sys.Date(),"%Y-%m-%d")
# rdata_file_name <- paste0("AvianInteractionData_L1_workspace", today_date_string, ".RData")
# save.image(file.path(file_paths$L1,rdata_file_name ))
```


## Exact Match on Common names

After applying exact and high-confidence fuzzy matches to scientific name/taxa,
those remaining unresolved species may have had a species name change but the
common name is preserved.  
Try exact matching on common names that are in Clement's. 

Exact matches will have fuzzy match score=1.0, but just less coding to go through
the whole matching process instead of direct match and then fuzzy match

```{r}
on_checklist_note <- "L0 species matched by common name to checklist (Clements 2024)"

# species_categories defined in checklist reading section above
checklist.common<- filter(checklist, category %in% species_categories)$common_name
```

Build the common name matching table

```{r}
# create list of names to match from unresolved rows
name_matches<- unresolved_species(int.raw.names)
# add new columns
name_matches$row_number <- unresolved_species_rows(int.raw.names)
name_matches$closest_name_match <- NA
name_matches$name_match_score <- NA
name_matches$scientific_name.checklist <- NA

# fill them in one row at a time
for(r in 1:nrow(name_matches)) {
  # optional, print the name we are on to gauge progress
  # print(scientific_name_matches[r,"scientific_name.edit"])
  
  raw_common_name<-name_matches[r,"common_name.raw"] 
  fuzzy_match <- find_closest_match_with_score(raw_common_name, checklist.common)
  
  # find the associated sci name, but only look at species records 
  # (some groups have the same common name for the whole genus)
  scientific_name_lookup <- as.character(filter(checklist, 
                                    common_name == fuzzy_match$match & category %in% species_categories ) %>%   
                          select(scientific_name))
  if(length(scientific_name_lookup) == 0) warning(paste(" no scientific name in checklist for ",fuzzy_match$match))
  else {
    name_matches[r,]$closest_name_match <- fuzzy_match$match
    name_matches[r,]$name_match_score <- fuzzy_match$score
    name_matches[r,]$scientific_name.checklist <- scientific_name_lookup 
  }
}
```

To view this table, run `View(name_matches)`

#### Example Matches

Considerations for setting the threshold (September 2025)

- *Northern Jancan* matches *Northern Jacana* 0.9458333	=> correct
- *Black-Winged Babbler*	matches *Black-chinned Babbler*	0.9190476 => incorrect 
- *Cackling/Canada Goose* matches *character(0)* 0.9099206 => incorrect (should be Branta hutchinsii/canadensis)

Conclusion setting threshold for match scores minimum 0.940 for high-confidence matches
which corrects a good number of species.  

```{r}
common_name_match_threshold <- 0.94
```

Update name edits using common name matches

```{r}
#| echo: false

high_confidence_common_matches <- filter(name_matches, name_match_score >= common_name_match_threshold)

# update by Row number. Not very R-like but works
# get the row numbers of names list already updated -> this process will not overwrite existing notes
# if following the notebook in order, these are mutually exclusive, but this ensures it could be 
# run out of order
rows_with_non_empty_notes <- which(!is.na(int.raw.names$edit_notes))

rows_to_update <- setdiff(high_confidence_common_matches$row_number, rows_with_non_empty_notes) # [!rows_high_confidence_match  %in% rows_with_non_empty_notes ]

print(paste("updating", length(rows_to_update), "rows"))

common_match_note <- "high confidence common name fuzzy match (Clements 2024) "

# make a copy (since matching takes a while)
int.raw.names.matches<- int.raw.names
# using saved row numbers, update scientific_name.edit and edit_note in bulk
int.raw.names.matches[high_confidence_common_matches$row_number,]$scientific_name.edit<- high_confidence_common_matches$scientific_name.checklist
# paste in the confidence level of these matches
int.raw.names.matches[high_confidence_common_matches$row_number,]$edit_notes <- paste( common_match_note, round(high_confidence_common_matches$name_match_score,3))
```

Remaining unresolved rows after high confidence scientific and common name matches:

```{r}
#| echo: false
#| output: true
print("unresolved rows in names table:")
print(length(unresolved_species_rows(int.raw.names.matches)))
print("unresolved unique scientific names ")
print(length(unique(unresolved_species(int.raw.names.matches)$scientific_name.raw)))

```

```{r}
#| echo: false
#| output: true

# after review, copy these resolve to the main names list
int.raw.names <- int.raw.names.matches
```

## Optional, Save the unresolved 

Save our current unresolved species after trying Fuzzy matching on Scientific name
and Common name, but prior to GBIF matching.  This is the basis for a table
aux_taxonomy_resolutions.csv

This maybe useful for comparison in future years

First, add a custom column with some notes
```{r}
int.raw.names.unresolved <- unresolved_species(int.raw.names)
int.raw.names.unresolved$row_number <- unresolved_species_rows(int.raw.names)
int.raw.names.unresolved$comments <- ""

## add a comment with the match score for common name
# first create the comment to use linked by row number we saved during that process
common_name_match <- filter(name_matches, row_number %in% int.raw.names.unresolved$row_number) %>%
    mutate(comment = paste("common name match: ", closest_name_match, round(name_match_score,2), scientific_name.checklist, ";")) %>% select(row_number, comment)
# update the comment, linking on row number
int.raw.names.unresolved<- mutate(int.raw.names.unresolved, comments = paste(comments, common_name_match[common_name_match$row_number == row_number, "comment"]))

## add comment with the match score for scientific name
# first create the comment to use linked by row number we saved during that process
scientific_name_match <- filter(scientific_name_matches, row_number %in% int.raw.names.unresolved$row_number ) %>%
  mutate(comment = paste("scientific name match: ", closest_scientific_name_match, round(scientific_name_match_score,2),";")) %>% 
  select(row_number, comment)
  
# add this match score to the comments in our unresolved list
#int.raw.names.unresolved
int.raw.names.unresolved<- mutate(int.raw.names.unresolved, 
           comments = paste(comments, scientific_name_match[scientific_name_match$row_number == row_number, "comment"]))

```

Optional: show this table of remaining unresolved to compare with current list

```{r}
# tibble(int.raw.names.unresolved)
```

Optional: Save CSV of our current unresolved species as the basis for a taxonomy resolution table.  

```{r}
# today_date_string <- format(Sys.Date(),"%Y-%m-%d")
# temp_file_name <- paste0("unresolved_L1", today_date_string, ".csv")
# write.csv(int.raw.names.unresolved, file=file.path(file_paths$L1,temp_file_name ), row.names = FALSE)
# print(paste("saved ", file.path(file_paths$L1,temp_file_name )))
```





## TaxaDB and GBIF Matching

taxadb package: Modified from this code:
https://docs.ropensci.org/taxadb/articles/intro.html 

In previous code, tried ITIS and COL to see if they were better than GBIF. 
GBIF has the fewest NA values, so we are sticking with it (it is the most comprehensive).

The GBIF database is frequently at odds with the Clements checklist.  Taxadb will
help find synonyms for a scientific name, but the result is often not on the checklist. 

However, it may be good at finding a handful of synonyms that _are_ on the checklist. 

This process takes the remaining unresolved names, finds the accepted GBIF name for them, 
and if that name is also on Clements, that is what will be assigned


#### Create a local GBIF database

Create a local gbif database in :
```{r}
taxadb::td_create("gbif")
```

Resolve scientific names for the scientific names, 
based on scientific_name.edit using GBIF

```{r}
int.gbif.names <- unresolved_species(int.raw.names) %>%
  mutate(
    scientific_id = taxadb::get_ids(scientific_name.edit, "gbif"),
    accepted_scientific_name = taxadb::get_names(scientific_id, "gbif")
  )
warnings()
```
**list join warnings**

For those entries that are ambiguous (\> 1 match of the edited species with gbif),
a warning will be issued. This loops through those warnings and prints 
the message that has the species in question so we can view and capture it.

```{r}

ws<- dplyr::last_dplyr_warnings()
print(paste("there are ", length(ws), "species have >1 identifier and need to be resolved w BOW"))

for( w in ws) print(w$parent)
```

IF there are duplicate matches (there will be warnings here)


####  TaxaDB/GBIF Edits review

of unresolved, how many had GBIF matches:

```{r}
#| echo: false
#| output: true

print(
  paste(
    nrow(filter(int.gbif.names, !is.na(scientific_id) )),
    "of",
    length(unresolved_species_rows(int.raw.names))
  )
)
```

How many of those gbif matches are actually on the checklist?

```{r}
#| echo: false
#| output: true

# checklist.species is defined above for other matching 
gbif_on_checklist <- filter(int.gbif.names, accepted_scientific_name %in% checklist.species$scientific_name )
print(nrow(gbif_on_checklist))
```

*TO-DO* 

we know it is suggesting Gallus gallus for chicken but that is currently not correct 
by checklist, so remove all of those 

```{r}
gbif_on_checklist <- filter(gbif_on_checklist, accepted_scientific_name != 'Gallus gallus')
print(nrow(gbif_on_checklist))
```


apply these changes from GBIF discovery

```{r}
# make a copy 
int.raw.names.gbif <- int.raw.names
count_rows_matched_via_gbif <- 0
# just use a loop through gbif matches to update, because sometime multiple rows match a sci name (due to sciname+common name combos)
for(r in 1:nrow(gbif_on_checklist)){
  names_row <- which(int.raw.names.gbif $scientific_name.raw == gbif_on_checklist[r,"scientific_name.raw"] & 
                int.raw.names.gbif$common_name.raw == gbif_on_checklist[r,"common_name.raw"])
  
  int.raw.names.gbif[names_row, ]$scientific_name.edit <- gbif_on_checklist[r,]$accepted_scientific_name
  gbif_note <- paste("GBIF: Match using taxadb/", gbif_on_checklist[r,]$scientific_id, " on checklist (Clements 2024)")
  
  int.raw.names.gbif[names_row, ]$edit_notes <- gbif_note
  count_rows_matched_via_gbif <- count_rows_matched_via_gbif + nrow( int.raw.names.gbif[names_row, ])
}

# save the count matched names for summary
count_names_matched_via_gbif <- nrow(gbif_on_checklist)
```

Optional Review outcome of applying GBIF
```{r}
# tibble(filter(int.raw.names.gbif, grepl("GBIF", edit_notes)))
```

After Review, save copy back to taxonomy table variable

```{r}
int.raw.names <-  int.raw.names.gbif 

```


Remaining species to resolve after fuzzy and gbif matching:

```{r}
#| echo: false
#| output: true
print("unresolved rows in names table:")
count_postGBIF_umatched_rows <- length(unresolved_species_rows(int.raw.names.gbif))
print(count_postGBIF_umatched_rows)
print("unresolved unique scientific names ")
count_postGBIF_umatched_scinames <- length(unique(unresolved_species(int.raw.names.gbif)$scientific_name.raw))
print(count_postGBIF_umatched_scinames)

```


## Scientific Name Changes: Low Confidence Matches on scientific name

**OPTIONAL:This is not normally necessary given so few unresolved species using current methods**

Extract scientific name matches with lower confidence (match_score <= 0.90) 
for those remaining species that haven't been resolved to see if
this may be helpful.  

*As of Fall 2025, there are so few left to match ( < 50 ) and low-confidence fuzzy matching has not be fruitful, and not helpful for 'unids', so it's unused.*

*The remaining resolutions are done manually (see below). This code is left here for reference*

```{r}
#| echo: false
# low_confidence_matches <- scientific_name_matches %>%
#   filter(
#     (scientific_name_match_score < 0.9 & !is.na(scientific_name_match_score)) 
#   ) %>% 
#   filter( # remove unid species from matches
#     !grepl(" sp\\.$", closest_scientific_name_match)
#   ) %>% 
#   filter(
#     scientific_name.raw %in% unresolved_species(int.raw.names)$scientific_name.raw
#   ) 
# 
#   # couldn't get Mutate to work, so here is a loop to add in the common name where it's found
#   low_confidence_matches$common_name.checklist <- NA
#   for(r in 1:nrow(low_confidence_matches)){
#     common_names <- filter(checklist,scientific_name == low_confidence_matches[r,"closest_scientific_name_match"] & category == "species")$common_name
#     if(length(common_names)==1) low_confidence_matches[r,"common_name.checklist"]<- common_names[1]
#   }
# 
# 
# length(unique(low_confidence_matches$scientific_name.raw))
```

**low-confidence match tallies:**

- previously ( early 2025): 69 out of 326 - checks out OK; 
- May 29, 2025: 65
- Sept 2025: 9 
- since Fall 2025, no longer used



## Unid : Genus-only species fixes

Open the Unid sp. (Genu-only) to work on:

Number of un-id names (genera only, not expected to Taxadb to match, just accounting):

```{r}
int.raw.names.unid <- filter(int.raw.names, grepl(" sp\\.$", scientific_name.edit))
print(paste("rows:", nrow(int.raw.names.unid)))
print(paste("unique:", length(unique(int.raw.names.unid$scientific_name.edit))))

```

**These unids have been added to the `aux_taxonomony_resolutions` file Feb 2026**


### Sumary Counts after 

<!-- *TO-DO: move this into a function so can be called at each step in workflow -->

*The goal of this step in the workflow is to resolve any taxa in the database to match
with an entry in Clement's checklist.  This summarizes how we are doing so far, prior to 
applying the 'manual' resolutions from `aux_taxonomy_resolutions` CSV*

Total number of name combinations (scientific+ common name) in database:

```{r}
#| echo: false
#| output: true

print(nrow(int.raw.names))

```

**number of full species that have found a resolution method (excluding unid, Genus only)**
```{r}
#| echo: false
#| output: true

print(nrow(unresolved_species(int.raw.names)))
```

number of all rows without notes (dispensation/edits) including Unid.:

```{r}
#| echo: false
#| output: true

print( nrow( filter(int.raw.names, is.na(edit_notes))))

```

To view the unresolved, run "tibble(unresolved_species(int.raw.names))"


```{r}
#| echo: false
#| output: true

# # optional checkpoint save

# today_date_string <- format(Sys.Date(),"%Y-%m-%d")
# rdata_file_name <- paste0("AvianInteractionData_L1_workspace", today_date_string, ".RData")
# f <- file.path(file_paths$L1,rdata_file_name )
# save.image(f)
# 
# print(paste("whole enviroment saved as ", f))

```

## Manual Taxanomic Edits

Not all automatic fixes were correct or some simply must be determined 
manually. These are collected in a CSV file with the same columns as our `int.raw.names`
data frame with columns

scientific_name.raw : name to match our name edits table 
common_name.raw : used primarily for reference when building the table
scientific_name.edit : the scientific name to apply in final output after review and decision
edit_notes : the reasoning why the scientific name was chosen to apply to the final output
comments : internal notes used to help build this table, not used in final output

These edits are apply on top of the 'edits' table built by automatic processes 
above. Some of these are 'overrides' where an entry applied automatically (e.g.
high confidence common name) was incorrect but easier to apply and correct here or
a typo that would be impossible for an automatic process to detect. 

There is code above to create a version of this table to start with if there is 
not such a table, based on the species that are not resolved so far but inserts 
the confidence of the matches in the comments field.  That may be used to 
create an additional table with new rows that can be copied/pasted into 
the main table with existing name resolution decisions

At the end of this process, a final name resolution table is output which includes
all of the name resolutions include those 

**TO-DO**

1. complete the entries in this file - PLZ completed 10/5/2025; LM confirmed 12/11/2025
3. read in the file and verify all rows are ok, filter out unfinished rows
4. apply the edits in the file regardless if there is already a note since 
   these may override the automatic resolutions
   
First step, check that the path to the manual resolutions file is named in a 
variable and that the file exists.  

If it does not this process will skip to the end leaving some species unresolved. 

```{r}
#| label: read_manual_resolutions
manual_taxonomic_resolutions <- data.frame()

if(file.exists(manual_taxonomic_resolutions_file_name)) {
  manual_taxonomic_resolutions <- read_csv(manual_taxonomic_resolutions_file_name, show_col_types = FALSE) 
} else {
  warning(paste("Manual taxonomic resolutions file not found", manual_taxonomic_resolutions_file_name))
  # couldn't find the file but keep going and check if we have an empty data frame
}

manual_taxonomic_resolutions <- filter(manual_taxonomic_resolutions, !is.na(scientific_name.edit))
print(paste("read in ", nrow(manual_taxonomic_resolutions), "manual name resolutions"))
```

**Manual Taxonomic Resolution Workflow: **

Goal is to accumulate a list of fixes and notes

1. review existing edits below
   use a lookup function to see if an edit (raw -> edit, common name) is still needed

1. Review the outputs of various steps of corrections
   The lower-confidence matches that may be skipped are a good source of 
How to to add an edit/adjustment to the list:

1. Preferred: add the entry in the file
1. call the `add_name_edits` function with arguments including the data frame
   used to collect these edits (`int.raw.names`, see above) directly

1. add to this list below with the arguments in the following order, which are 
   added one by one calling that `add_name_edits` function
  
   - original scientific_name (raw) as appears in interaction file, 
   - editing notes, dispensation or why an edit is needed or d, 
   - (optional) scientific_name.edit IF the name is changed
   - (optional) common_name.edit if the name was changed 



```{r}
#| label: apply_manual_resolutions

int.raw.names.manual <- int.raw.names
for(r in 1:nrow(manual_taxonomic_resolutions)){
  manual.edit = manual_taxonomic_resolutions[r,]
  matching_rows = which(int.raw.names.manual$scientific_name.raw == manual.edit$scientific_name.raw)
  # TODO check if r is a real vector of integers
  if(length(matching_rows) == 0){
    # manual edited raw species is not in L0 data, may be left over,
    not_found_msg <- paste(manual.edit$scientific_name.raw, ": this raw /L0 species in manual edits table no longer found in L0 data" )
    warning(not_found_msg)
  } else {
    int.raw.names.manual[matching_rows,]$scientific_name.edit <- manual.edit$scientific_name.edit
    int.raw.names.manual[matching_rows,]$edit_notes <- manual.edit$edit_notes
  }

}

```

Remaining species to manual resolve after fuzzy and gbif matching:

```{r}
#| echo: false
#| output: true
print("unresolved rows in names table:")
print(length(unresolved_species_rows(int.raw.names.manual)))
print("unresolved unique scientific names ")
print(length(unique(unresolved_species(int.raw.names.manual)$scientific_name.raw)))

```

To view, run View(unresolved_species(int.raw.names.manual))

If updating the CSV, re-run from chunk `read_manual_resolutions`

After review, save this list to the main name resolution data frame

```{r}
int.raw.names <- int.raw.names.manual

```

review data frame to make decisions

```{r}
# make additional edits using int.raw.names <- add_name_edits(...)
# re-run the block subset_check

```


### Subspecies Manual Resolution

check if there are any subspecies remaining that have interactions

for those, decide what to do with them if they exist, potentially publish in the 
database


```{r}
#| eval: false


```


### Work through hybrids 

```

Hybrids to standardize: 
  
  TO-DO: keep them in the database, but fix only the following 

  Hybrid Bullock's Oriole X Baltimore = Oriole Icterus bullockii x galbula
  Hybrid Western Gull X Glaucous-Winged Gull = Larus occidentalis x glaucescens
  Hybrid Barnacle X Bar-Headed Goose = Branta leucopsis x anser indicus

dplyr::filter(int.raw, taxa1_scientific %in% c("Branta leucopsis x"))
# NO HYBRIDS IN CLEANED DATA NOW
dplyr::filter(int.raw, taxa1_scientific %in% c("Branta leucopsis x Anser indicus"))
dplyr::filter(int.raw, taxa2_scientific %in% c("Branta leucopsis x Anser indicus"))
# dplyr::filter(unresolved.gbif, scientific_name.edit %in% c("Branta leucopsis x anser indicus"))
# dplyr::filter(unresolved.gbif, scientific_name.edit %in% c("Branta leucopsis x anser indicus"))
# fixed_names1$scientific_name[fixed_names1$scientific_name.edit == "Branta leucopsis x anser indicus"] <- "Branta leucopsis x Anser indicus"
```



## Merge and Save data

Update resolved common names use the Clements/eBird 2024 checklist to assign common names (by referencing the official scientific name; all fixed above)

```{r}
# mutate is terrible at joining things and I only want to join common name
for(r in 1:nrow(int.raw.names)){
  common_name.checklist <- checklist[checklist$scientific_name == int.raw.names[r,]$scientific_name.edit,]$common_name
  if(length(common_name.checklist) > 0 ) { # match 
    int.raw.names[r,]$common_name.edit <- common_name.checklist
  }
}

```

Update the scientific names in the interactions data with our table as much 
as possible: 

```{r}
# make a copy of the database data frame so we can re-run if necessary
int.raw.resolved <- int.raw

updated_rows <- c()

for(r in 1:nrow(int.raw.names)){
  # if no resolution discovered, skip (for now - ultimately all are updated)
  if(is.na(int.raw.names[r,"scientific_name.edit"])){
    next
  }
  
  # extract values from data frame for readability
  scientific_name.raw <- int.raw.names[r,]$scientific_name.raw
  scientific_name.edit <- int.raw.names[r,]$scientific_name.edit
  common_name.edit <- int.raw.names[r,]$common_name.edit
  
  # update sp1s and sp2s
  # use matches from original dataframe but update our resolved data frame 
  # since the rows are the same order and number 
  # this allows for matching on the raw sci name for common name updates, 
  # even after it's been updated
  
  updated_sp1_rows = which(int.raw$taxa1_scientific == scientific_name.raw)
  if(length(updated_sp1_rows) > 0){
    int.raw.resolved[updated_sp1_rows, ]$taxa1_scientific <- scientific_name.edit
    int.raw.resolved[updated_sp1_rows, ]$taxa1_common <- common_name.edit
    updated_rows <- union(updated_rows, updated_sp1_rows)
  }

  updated_sp2_rows = which(int.raw$taxa2_scientific == scientific_name.raw)
  if(length(updated_sp2_rows) > 0){
    int.raw.resolved[updated_sp2_rows, ]$taxa2_scientific <- scientific_name.edit
    int.raw.resolved[updated_sp2_rows,  ]$taxa2_common <- common_name.edit
    updated_rows <- union(updated_rows, updated_sp2_rows)
  }
  
}

print(length(updated_rows))
```

Review the updated database using `View(int.raw.resolved)`  The original is still in `int.raw`

Some manual updates for Northwestern Crow (common name doesn't exist anymore but the data still contain the subspecies; re-assign to American Crow)
```{r}

int.raw.resolved$taxa1_scientific[int.raw.resolved$taxa1_scientific == "Corvus brachyrhynchos caurinus"] <- "Corvus brachyrhynchos"

int.raw.resolved$taxa1_common[int.raw.resolved$taxa1_scientific == "Corvus brachyrhynchos"] <- "American Crow"

int.raw.resolved$taxa2_scientific[int.raw.resolved$taxa2_scientific == "Corvus brachyrhynchos caurinus"] <- "Corvus brachyrhynchos"

int.raw.resolved$taxa2_common[int.raw.resolved$taxa2_scientific == "Corvus brachyrhynchos"] <- "American Crow"
```

```{r}
# after review, copy int.raw.resolved to main variable 

int.raw <- int.raw.resolved

```

## Record the resolved taxomy in database

*Create new column for the Clements string for the scientific name*

This should mostly be the same as the taxa1_scientific and taxa2_scientific, except for a few unid. where our level of taxonomic resolution is higher than any option in the Clements species list

```{r}
# Combine both scientific name columns from the interaction data
# (taxa1 = species A, taxa2 = species B), then keep only unique names
unique_taxa_list <- unique(c(int.raw$taxa1_scientific, int.raw$taxa2_scientific))

# Extract and store only the names that do NOT match Clements taxonomy
# These are the mismatches that need correction/review
mismatch <- unique_taxa_list[!(unique_taxa_list %in% checklist$scientific_name)]

# For taxa already labeled according to Clements, copy taxa1_scientific into taxa1_clements
# For taxa not labeled according to Clements, put NA
int.raw$taxa1_clements <- ifelse(
  int.raw$taxa1_scientific %in% mismatch, 
  NA, 
  int.raw$taxa1_scientific
)

int.raw$taxa2_clements <- ifelse(
  int.raw$taxa2_scientific %in% mismatch, 
  NA, 
  int.raw$taxa2_scientific
)

# make a clean lookup
lookup <- clements_mismatches %>%
  select(old_name, new_name)

# Join in new Clements names where needed
int.raw <- int.raw %>%
  left_join(lookup, by = c("taxa1_scientific" = "old_name")) %>%
  rename(new1 = new_name) %>%
  left_join(lookup, by = c("taxa2_scientific" = "old_name")) %>%
  rename(new2 = new_name) %>%
  mutate(
    taxa1_clements = if_else(!is.na(new1), new1, taxa1_clements),
    taxa2_clements = if_else(!is.na(new2), new2, taxa2_clements)
  ) %>%
  select(-new1, -new2)

# Identify any remaining mismatches 
remaining_mismatch <- unique(c(int.raw$taxa1_scientific[is.na(int.raw$taxa1_clements)], 
                               int.raw$taxa2_scientific[is.na(int.raw$taxa2_clements)]))


# Save list of new mismatches if relevant (if not, then delete the previous list)
# These will be copied into the clements_mismatches csv and manually rectified
if(length(remaining_mismatch) > 0){
  write.csv(remaining_mismatch, here::here("./data/L1/aux_mismatches_with_clements_TOFIXMANUALLY.csv"), row.names=FALSE)
}else(file.remove(here::here("./data/L1/aux_mismatches_with_clements_TOFIXMANUALLY.csv")))

```


## Print Database Summary

```{r}
#| label: database_summary
#| echo: false
#| output: true

print("unique scientific names = number of unique species in this database")
print(length(union(int.raw$taxa1_scientific, int.raw$taxa2_scientific)))


print("unique types of interactions")
print(length(unique(int.raw$interaction))) #36 (37 with manual fixes)

print("unique pairwise interactions") #19417 (non-manual fixes); 19403 with manual fixes

# if you want to ignore the order of species, we will need to create pairs and sort them
# some sample code saved here to consider: 
# dat <- data.frame(c('a','b','c'), c('b','a','d'), stringsAsFactors=F );
# colnames(dat) <- c('s1','s2');
# x <- unique(as.data.frame(t(apply(dat, 1, sort ))));
# rownames(x) <- 1:nrow(x);
# x;
print(length(unique(paste(int.raw$taxa1_scientific, int.raw$taxa2_scientific))))

# edit to report number of unique rows by DATA$taxa1_scientific, DATA$taxa2_scientific, DATA$interaction

```


## Save Results to CSV

```{r}
#| label: save_output
#| echo: false
#| output: true

# set these variables here for flexibility if the workflow changes 
interaction_data <- int.raw
taxonomic_resolutions<- int.raw.names

# AvianInteractionData_L1_file and AvianInteractionData_L1_taxonomy_table is set at the top of this file

f <- file.path(file_paths$L1, AvianInteractionData_L1_file )
write_csv(interaction_data, f)
print(paste("interaction database saved to ", f))

f<- file.path(file_paths$L1, AvianInteractionData_L1_taxonomy_table_file )
write_csv(taxonomic_resolutions, f)
print(paste("taxonomic_resolutions_file saved to ", f))

```

## Summary Counts


These are counts of taxa from the full list extracted from the L0, cleaned database. 
Each "taxa" on this list is a unique combination of (Scientific taxa + Common name), where 
sometimes the common name entred may vary depending on the original source, creating unique combinations


|taxa  |  count | description |
+------+--------+-------------+
| All  | 435    | number of unique taxa from cleaned data that were not in Clement's list |
|      |        | matches found from high-confidence fuzzy search on Scientific Name |
|      |        |




Throughout the workflow, all species names were aligned with the Clements/eBird 2024 checklist (REF) for taxonomic harmonization, and any discrepancies were changed in code (Fig. 2 “Clean database”). To harmonize taxonomy for scientific names that did not exactly match species on the Clements/eBird checklist (425 taxa), we ran through a series of semi-automated steps to harmonize the taxonomy. First, we used fuzzy text search on scientific names with a threshold of 0.95 to automatically correct potential typos. For the remaining unresolved names, we used the rGBIF and taxaDB packages to identify known synonyms, resolving XXX taxa. Finally, for the remaining discrepancies (72 taxa), we manually evaluated the scientific names, referring to the Clements/eBird 2024 checklist (REF), Avibase (REF), the 2024 BBS species list (REF), and BOW (REF) to create a manual taxonomic cross-walk table (REF for this manual edit table). In code, all taxonomy changes were added to a master crosswalk table (REF FOR TABLE) and reviewed prior to altering the AvianMetaNetwork data.  

```{r}
# first pass at printing the summary counts (ultimately should go into markdown table)

print(paste("count_L0_total_unique_species_unedited ", count_L0_total_unique_species_unedited))
print(paste("count_L0_rows_scinames_already_on_checklist ", count_L0_rows_scinames_already_on_checklist))
print(paste("count_L0_species_scinames_already_on_checklist ", count_L0_rows_scinames_already_on_checklist))

print(paste("count_scinames_needing_to_be_matched after checklist match", count_scinames_needing_to_be_matched ))

print(paste("count_rows_needing_matching_after_clements_match ", count_rows_needing_matching_after_clements_match ))


print(paste("count_scinames_unresolved_prior_to_sciname_fuzzy_matching", count_scinames_unresolved_prior_to_sciname_fuzzy_matching))
print(paste("high_confidence fuzzy match threshold", high_confidence_threshold))

print(paste("count_names_matched_via_gbif ", count_names_matched_via_gbif))
print(paste("count_rows_matched_via_gbif ",  count_rows_matched_via_gbif)) 
print(paste("count_postGBIF_umatched_rows ", count_postGBIF_umatched_rows))
print(paste("count_postGBIF_umatched_scinames", count_postGBIF_umatched_scinames))



```

## Subset to the checklist 

**Notes about next steps for subsetting**

This should go into a different script/notebook

Analysis Paper: 

1 sp1 and sp2 on checklist (canada CONUS)
Data Paper Goal: 

1. Sp1 or sp2 needs to be on Checklist (canada alaska conus checklist)
2. Final check:which rows are not (sp1 AND sp2 on Canada Alaska Conus checklist)?
    how many?
    which of these are NOT on full clements list
    how many? 
    review data
    check against all clements
    assume names are fixed via gbif/taxa db
    
    if descions are made add name edits to be added to the int.raw.names list above



```{r}
#| label: subset_check
#| echo: false

# rcode for above, prints a data frame of non-checklist rows

```

---

```{r}
sessionInfo()
```
