==> R/L1/5_subset_network.qmd <==
---
title: "L1 post-taxonomic harmonization subset to Canada/AK/CONUS"
author: 
  - Phoebe L. Zarnetske
  - Patrick Bills
format:
  html:
    toc: true
    toc-depth: 3
    embed-resources: true
editor: source
date: 11/06/2025
date-format: iso
---

## Avian Interaction Pairs Data: L1 to regional subset for Canada/Alaska/CONUS

### About

- **PROJECT:** AvianMetaNet (CONUS/Canada/AK) and MetaNetwork
- **COLLABORATORS:** Vincent Miele, Stephane Dray
- **Run Date:** `r Sys.Date()`

### Data inputs:

**AvianMetaNet L0**

Raw Data, combined (stitched) and with content
cleaning but taxonomic names as entered (older names, typos, etc)

*Set the interactions L1 database file to read from here:*


( output From `R/L1/AvianInteractionData_L1.R` )

```{r}
# Clear all existing data
rm(list=ls())

AvianInteractionData_L1_file <-  "amn_all.csv"

```

**Main Checklists:**

Curated Species checklist: 

- Current selected species list for Canada/Alaska/CONUS =
  combined checklist based on Clements/eBird v2024, Avibase v8.17, BBS v2024:
 ( output via Working repo, from `R/L1/3_subset_species_lists.R` ) = spp_joint_cac_colsubset.csv



### DATA OUTPUT:

L1 Canada, Alaska, CONUS species from L1 AvianMetaNet 
  (`amn_all.csv`); saved as:

```{r}
AvianInteractionData_CanadaAKCONUS_L1_file <- "amn_cac.csv"
EDI_file <- "AvianMetaNet_InteractionData_EDI.csv"
AvianInteractionData_CanadaAKCONUS_breeding_file <- "amn_cac_breeding.csv"
AvianInteractionData_L1_not_CanadaAKCONUS_file <- 
  "amn_notcac.csv"
today_date_string <- format(Sys.Date(),"%Y-%m-%d")

```


### Next script to run: 

- depends on project: 
  - for AvianMetaNet North America (Data Paper): a summary and 
  figure-generating script (to be written)
  - for Avian MetaNetwork Analysis: https://github.com/SpaCE-Lab-MSU/avian-meta-network/tree/main/R/L2


### NOTES

This script is used to subset out only the species that occur in CONUS+Alaska+Canada plus some minor cleaning for the North American subset publication.  


## Data Set-up

File paths used for this run:

```{r}
#| echo: false

# functions used by all 
source(here::here('R/lib/shared_functions.R'))
# functions for this process to reduce this script size
# and load libraries used here
source(here::here('R/lib/taxonomy_functions.R'))

file_paths <- get_file_paths()

print(file_paths)
```

*Note that the Checklist folder above is in L0 which contains raw species lists, 
but the Canada/Alaska/CONUS L1 Checklist is the one to use here because it is 
built from multiple L0 lists*


### Setup: Checklists

#### Curated List

October 2025: new curated list curated by PLZ using Clements, AviBase, BBS etc. 
See `R/L1/3_subset_species_lists.R`

```{r}

current_species_list_filename <- "spp_joint_cac_colsubset.csv"
splist_filepath <- file.path(file_paths$CHECKLIST_L1, current_species_list_filename )
splist <- read.csv(here::here(splist_filepath))
print(paste("Read in ", nrow(splist), "entries from checklist file", splist_filepath))

```



#### Interactions Data

To use a different or test interaction files, edit this L1 filename:

```{r}
intxns.db.file <- file.path(file_paths$L1,AvianInteractionData_L1_file)
intxns.db <-read.csv(here::here(intxns.db.file))
print(paste("read in ", nrow(intxns.db), "records from ", intxns.db.file))
```
#### Remove columns only relevant for non-NA species 

```{r}
intxns.db <- intxns.db %>% dplyr::select(-DatabaseSearchURL)
```


- get all interactions where either sp1 or sp2 is a species on the Canada/AK/CONUS list 
- subset out a few relevant columns in the splist for use with BBS network and spatial analysis

Use scientific_name_clements2024 as the match for names.
```{r}
length(unique(splist$scientific_name_clements2024)) #777
dim(splist) #785
dim(intxns.db)  #31040 (30776 in manual fix version)
# Some duplicate rows; these are just the unids. so we can ignore them
duplicate_indices <- duplicated(splist$scientific_name_clements2024)
duplicate_clements <- splist[duplicate_indices, ]
duplicate_clements

intxns.db.subset <- intxns.db[intxns.db$taxa1_scientific %in% splist$scientific_name_clements2024 | intxns.db$taxa2_scientific %in% splist$scientific_name_clements2024,]
dim(intxns.db.subset)  #26200 (25958 in manual fix version)

intxns.db.non.subset <- intxns.db[!(intxns.db$taxa1_scientific %in% splist$scientific_name_clements2024 | intxns.db$taxa2_scientific %in% splist$scientific_name_clements2024),]
dim(intxns.db.non.subset) #4818 in manual fix version
# Check that these equate to dim(intxns.db)
dim(intxns.db.non.subset)+dim(intxns.db.subset)
dim(intxns.db)  #30776 in manual fix version

usp1<-(unique(intxns.db.subset$taxa1_scientific))
usp2<-(unique(intxns.db.subset$taxa2_scientific))
usp1.2<-append(usp1,usp2)
length(unique(usp1.2)) #2137 unique species in the interaction data

# 2122 in manual fix version
# Create a simplified Clements/eBird BBS AOU combo splist:
splist<-splist %>%
      select(scientific_name_clements2024.combo, common_name_clements2024.combo,
             AOU_bbs2024.combo, scientific_name_clements2024, common_name_clements2024, 
             scientific_name_bbs2024, AOU_bbs2024, in_bbs2024)

```
### Remove entries with no equivalent Clements name for a taxa
## (E.g., Galliform sp.)

```{r}
intxns.db.subset <- intxns.db.subset[
  !is.na(intxns.db.subset$taxa1_clements) &
  !is.na(intxns.db.subset$taxa2_clements),
]
```


### Flip all non-CAC spp1 and CAC spp2 combos so that spp1 is always a CAC spp 

```{r}
cac_spp1 <- intxns.db.subset[intxns.db.subset$taxa1_scientific %in% unique(splist$scientific_name_clements2024),]
nocac_spp1 <- intxns.db.subset[!(intxns.db.subset$taxa1_scientific %in% unique(splist$scientific_name_clements2024)),]

nocac_spp1_flipped <- nocac_spp1 %>% 
  rename(
    taxa2_common = taxa1_common, 
    taxa1_common = taxa2_common, 
    taxa2_scientific = taxa1_scientific, 
    taxa1_scientific = taxa2_scientific, 
    effect_tx1_on_tx2 = effect_tx2_on_tx1, 
    effect_tx2_on_tx1 = effect_tx1_on_tx2
  ) %>% 
  relocate(colnames(nocac_spp1))

# Check to make sure all spp1 are now CAC species (value should be 0) 
length(nocac_spp1_flipped$taxa1_scientific[!(nocac_spp1_flipped$taxa1_scientific %in% unique(splist$scientific_name_clements2024))])

# Recombine 
intxns.db.subset.new <- rbind(cac_spp1, nocac_spp1_flipped)

network_spp <- unique(c(intxns.db.subset.new$taxa1_scientific, intxns.db.subset.new$taxa2_scientific))

# Check to make sure network contains all CAC species 
missing_spp <- setdiff(
  unique(splist$scientific_name_clements2024),
  network_spp
)

all(unique(splist$scientific_name_clements2024) %in% network_spp)

```


Adjust nonbreeding season column @billspat note to move this to the L0 cleaning

```{r}
# List of characters to re-assign
old1 <- c("YSE", "YES", "Yes", "possibly", "potential", "yes")
new1 <- "nonbreeding"
# Re-assign values 
intxns.db.subset.new$nonbreedingseason <- ifelse(intxns.db.subset.new$nonbreedingseason %in% old1, new1, intxns.db.subset.new$nonbreedingseason)

old2 <- "breeding and nonbreeding"
new2 <- "breeding"
# Re-assign values 
intxns.db.subset.new$nonbreedingseason <- ifelse(intxns.db.subset.new$nonbreedingseason %in% old2, new2, intxns.db.subset.new$nonbreedingseason)

# Replace NA in 'nonbreedingseason' with "breeding"
intxns.db.subset.new$nonbreedingseason[is.na(intxns.db.subset.new$nonbreedingseason)] <- "breeding"

# Rename the column to 'season'
intxns.db.subset.new <- intxns.db.subset.new %>%
  rename(season = nonbreedingseason)

```



Assign all competition- to competition for this version (only a few of these exist; they will be kept specific in v2 of the data)

```{r}


## COMPETITION 

# List of characters to re-assign
old_comp <- c("competition-foraging","competition-nest site","competition-territory")
new_comp <- "competition"
# Re-assign values 
intxns.db.subset.new$interaction <- ifelse(intxns.db.subset.new$interaction %in% old_comp, 
                                       new_comp, 
                                       intxns.db.subset.new$interaction)


## KLEPTOPARASITISM

# List of characters to re-assign
old_klepto <- c("kleptoparasitism-food","kleptoparasitism-nest material")
new_klepto <- "kleptoparasitism"
# Re-assign values 
intxns.db.subset.new$interaction <- ifelse(intxns.db.subset.new$interaction %in% old_klepto, 
                                       new_klepto, 
                                       intxns.db.subset.new$interaction)



## FACILITATION

# List of characters to re-assign
old_facil<- c("facilitation-creching","facilitation-distress calls")
new_facil <- "facilitation"
# Re-assign values 
intxns.db.subset.new$interaction <- ifelse(intxns.db.subset.new$interaction %in% old_facil, 
                                       new_facil, 
                                       intxns.db.subset.new$interaction)


# List of characters to re-assign
old_facil_sub<- c("facilitation-feeding", "facilitation-foraging")
new_facil_sub <- "facilitation-feeding/foraging"
# Re-assign values 
intxns.db.subset.new$interaction <- ifelse(intxns.db.subset.new$interaction %in% old_facil_sub, 
                                       new_facil_sub, 
                                       intxns.db.subset.new$interaction)



## PREDATION

# List of characters to re-assign
old_pred<- c("nest predation")
new_pred <- "predation"
# Re-assign values 
intxns.db.subset.new$interaction <- ifelse(intxns.db.subset.new$interaction %in% old_pred, 
                                       new_pred, 
                                       intxns.db.subset.new$interaction)


## BROOD PARASITISM

# List of characters to re-assign
old_para<- c("nest parasitism")
new_para <- "brood parasitism"
# Re-assign values 
intxns.db.subset.new$interaction <- ifelse(intxns.db.subset.new$interaction %in% old_para, 
                                       new_para, 
                                       intxns.db.subset.new$interaction)

## MOBBING

# List of characters to re-assign
old_mob<- c("shared scolding")
new_mob <- "mobbing"
# Re-assign values 
intxns.db.subset.new$interaction <- ifelse(intxns.db.subset.new$interaction %in% old_mob, 
                                       new_mob, 
                                       intxns.db.subset.new$interaction)


# Generate summary table of interactions by type 
num_intxn_types <- intxns.db.subset.new %>% group_by(interaction) %>% summarize(n=n()) %>% arrange(-n) 
knitr::kable(num_intxn_types)
```


Create summary information in a table
```{r}
unique_counts_multi <- purrr::map_dfr(
      .x = c("interaction", "season"), 
      .f = ~ intxns.db.subset.new %>%
        count(!!sym(.x)) %>%
        mutate(Column = .x) %>%
        rename(Value = !!sym(.x), Count = n) %>%
        select(Column, Value, Count)
    )
# Nicer format
    knitr::kable(unique_counts_multi,
          caption = "Breeding and Non-breeding Data: Summaries",
          col.names = c("Column", "Value", "Count"),
          format = "markdown")
```

Version with only breeding season interactions:

```{r}
# Subset rows where 'season' is "breeding"
intxns.db.subset.new.breeding <- intxns.db.subset.new %>%
  filter(season == "breeding")
dim(intxns.db.subset.new.breeding) #23164 (22969 in fixed manual version)
dim(intxns.db.subset.new) #26200 (25958 in fixed manual version)

# Summary table
unique_counts_multi.b <- purrr::map_dfr(
      .x = c("interaction", "season"), 
      .f = ~ intxns.db.subset.new.breeding %>%
        count(!!sym(.x)) %>%
        mutate(Column = .x) %>%
        rename(Value = !!sym(.x), Count = n) %>%
        select(Column, Value, Count)
    )
# Nicer format
    knitr::kable(unique_counts_multi.b,
          caption = "Breeding Season Data: Summaries",
          col.names = c("Column", "Value", "Count"),
          format = "markdown")

```

Create subset file for archival with EDI that does not include internal columns:  

```{r}

intxns.db.subset.edi <- intxns.db.subset.new %>% 
  dplyr::select(
    -recorder, 
    -uncertain_interaction,
    -n_studies,
    -name_changes,
    -entry_file,
    -source,
    -other_species1
  ) %>% 
  relocate(
    taxa1_common, 
    taxa1_scientific,
    taxa1_clements,
    taxa2_common,
    taxa2_scientific,
    taxa2_clements,
    interaction
  )

```


Save the file with the file name defined in the OUTPUT section above
this has the current date embedded in it

```{r}

f <- file.path(file_paths$L1, AvianInteractionData_CanadaAKCONUS_L1_file )
write_csv(intxns.db.subset.new, f)
print(f)


fedi <- file.path(file_paths$L1, EDI_file)
write_csv(intxns.db.subset.edi, fedi)
print(fedi)

fb <- file.path(file_paths$L1, AvianInteractionData_CanadaAKCONUS_breeding_file )
write_csv(intxns.db.subset.new.breeding, fb)
print(fb)

# THE NON-CAC SUBSET DOES NOT CURRENTLY GET UPDATED IN THE CODE WITH THE 
# GENERALIZED INTERACTION NAMES 
# fx <- file.path(file_paths$L1, AvianInteractionData_L1_not_CanadaAKCONUS_file )
# write_csv(intxns.db.non.subset, fx)
# print(fx)

```
